Moved permissions to own table.
Permissions for each API_KEY are now saved in their own table (ui_uihk_rest_perm) instead of
a VARCHAR(60000) column inside the ui_uihk_rest_keys table itself.
The benefit of this is:
 * The amount of permissions was limited by the maximum size of the VARCHAR, which seems
to wary between platforms and/or database implementations.
Using a separate table removes this (minor) restriction.
* This should improve the database layout ("Database Normalization") as this can now potentially
make use of the relational principle of SQL Databases