Fix database error and add phpdoc
When the token was generated a second time, a sql error occured, because of the primary key contstraint.
Now, it checks first, if the token with the according user id exists already and if so it will be returned,
instead of a new one.

In addition, phpdoc comments are added.